Refactor wavefront parser
