+++ sources/parser/wavefront.c
@@ -6,7 +6,7 @@
 /*   By: ppaulo-d <ppaulo-d@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/06/11 09:53:45 by ppaulo-d          #+#    #+#             */
-/*   Updated: 2023/06/21 13:13:26 by ppaulo-d         ###   ########.fr       */
+/*   Updated: 2023/06/21 16:48:02 by ppaulo-d         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,35 +21,6 @@ static	int	line_count(t_bool increment)
 	return (line_count);
 }
 
-static void	add_vertex(t_parser *parser, t_p3d vertex)
-{
-	parser->vertices = ft_realloc(parser->vertices,
-		sizeof(vertex) * parser->vtx_count, sizeof(vertex)
-			* (parser->vtx_count + 1));
-	parser->vertices[parser->vtx_count] = vertex;
-	parser->vtx_count++;
-}
-
-// note: this function will be removed when parse_coordinate is implemented
-static	t_bool	parse_vertex(t_parser *parser, char *line)
-{
-	char	**coordinates;
-	t_p3d	vertex;
-
-	++line;
-	line = jump_spaces(line);
-	if (check_coordinate(line) == false)
-		return (false);
-	coordinates = ft_split(line, ' ');
-	vertex.x = ft_atof(coordinates[0]);
-	vertex.y = ft_atof(coordinates[1]);
-	vertex.z = ft_atof(coordinates[2]);
-	vertex.w = 1;
-	add_vertex(parser, vertex);
-	ft_clean_array((void **) coordinates, free);
-	return (true);
-}
-
 static t_bool	parser_obj_line(t_parser *parser, char *line)
 {
 	line = jump_spaces(line);
@@ -60,7 +31,7 @@ static t_bool	parser_obj_line(t_parser *parser, char *line)
 	else if (ft_strncmp(line, "vt", 2) == 0)
 		return (true);
 	else if (ft_strncmp(line, "v", 1) == 0)
-		return (parse_vertex(parser, line));
+		return (parse_vertices(parser, line));
 	else if (ft_strncmp(line, "f", 1) == 0)
 		return (parse_faces(parser, line));
 	else if (ft_strncmp(line, "g", 1) == 0)
