+++ sources/parser/wavefront_parse_normals.c
